refactor: ChatRoom 컴포넌트 분리
